+++ pymongo/settings.py
@@ -20,9 +20,9 @@ from bson.objectid import ObjectId
 from pymongo import common, monitor, pool
 from pymongo.common import LOCAL_THRESHOLD_MS, SERVER_SELECTION_TIMEOUT
 from pymongo.errors import ConfigurationError
-from pymongo.topology_description import TOPOLOGY_TYPE
 from pymongo.pool import PoolOptions
 from pymongo.server_description import ServerDescription
+from pymongo.topology_description import TOPOLOGY_TYPE
 
 
 class TopologySettings(object):
@@ -36,7 +36,8 @@ class TopologySettings(object):
                  local_threshold_ms=LOCAL_THRESHOLD_MS,
                  server_selection_timeout=SERVER_SELECTION_TIMEOUT,
                  heartbeat_frequency=common.HEARTBEAT_FREQUENCY,
-                 server_selector=None):
+                 server_selector=None,
+                 fqdn=None):
         """Represent MongoClient's configuration.
 
         Take a list of (host, port) pairs and optional replica set name.
@@ -55,6 +56,7 @@ class TopologySettings(object):
         self._local_threshold_ms = local_threshold_ms
         self._server_selection_timeout = server_selection_timeout
         self._server_selector = server_selector
+        self._fqdn = fqdn
         self._heartbeat_frequency = heartbeat_frequency
         self._direct = (len(self._seeds) == 1 and not replica_set_name)
         self._topology_id = ObjectId()
@@ -99,6 +101,10 @@ class TopologySettings(object):
     @property
     def heartbeat_frequency(self):
         return self._heartbeat_frequency
+
+    @property
+    def fqdn(self):
+        return self._fqdn
 
     @property
     def direct(self):